Improve and implement textures

+++ RayTracerRedone/tracer/scene/textures/Texture.h
@@ -10,7 +10,7 @@
 class Texture {
 public:
     Texture()=default;
-    [[nodiscard]] virtual ColorVec value(double u, double v, const Point3& p) const = 0;
+    [[nodiscard]] virtual ColorVec value(float u, float v, const Point3& p) const = 0;
 };
